Patched menus

+++ cogs/menus.py
@@ -12,8 +12,9 @@ class MenuManager(commands.Cog):
     # Deletes menus after a certain time.
     async def delete_after(self, timeout: int, menu_id):
         await asyncio.sleep(timeout)
-        await self.active_menus[menu_id].remove()
-        del self.active_menus[menu_id]
+        if menu_id in self.active_menus:
+            await self.active_menus[menu_id].remove()
+            del self.active_menus[menu_id]
 
     async def new_menu(self, ctx: commands.Context, pages: list, cur_page: int = 0, timeout: int = 60):
         menu = Menu(ctx, pages, cur_page, timeout)
@@ -27,11 +28,13 @@ class MenuManager(commands.Cog):
             return
         if payload.message_id in self.active_menus:
             if payload.emoji.name == '◀':
-                await self.active_menus[payload.message_id].previous_page()
+                await self.active_menus[payload.message_id].previous_page(payload.user_id)
             elif payload.emoji.name == '⏹':
-                await self.delete_after(0, payload.message_id)
+                if payload.message_id in self.active_menus \
+                        and self.active_menus[payload.message_id].owner == payload.user_id:
+                    await self.delete_after(0, payload.message_id)
             elif payload.emoji.name == '▶':
-                await self.active_menus[payload.message_id].next_page()
+                await self.active_menus[payload.message_id].next_page(payload.user_id)
 
     @commands.Cog.listener()
     async def on_raw_reaction_remove(self, payload: discord.RawReactionActionEvent):
@@ -39,11 +42,13 @@ class MenuManager(commands.Cog):
             return
         if payload.message_id in self.active_menus:
             if payload.emoji.name == '◀':
-                await self.active_menus[payload.message_id].previous_page()
+                await self.active_menus[payload.message_id].previous_page(payload.user_id)
             elif payload.emoji.name == '⏹':
-                await self.delete_after(0, payload.message_id)
+                if payload.message_id in self.active_menus \
+                        and self.active_menus[payload.message_id].owner == payload.user_id:
+                    await self.delete_after(0, payload.message_id)
             elif payload.emoji.name == '▶':
-                await self.active_menus[payload.message_id].next_page()
+                await self.active_menus[payload.message_id].next_page(payload.user_id)
 
 
 class Menu:
@@ -55,29 +60,31 @@ class Menu:
         self.message = None
         self.id = ctx.message.id
         self.cur_page = cur_page
+        self.owner = ctx.author.id
 
     async def open(self):
         self.message = await self.ctx.send(embed=self.pages[self.cur_page])
-        print(self.message)
         await self.message.add_reaction('◀')
         await self.message.add_reaction('⏹')
         await self.message.add_reaction('▶')
 
-    async def next_page(self):
-        if self.cur_page < len(self.pages) - 1:
+    async def next_page(self, user_id):
+        if self.cur_page < len(self.pages) - 1 and user_id == self.owner:
             self.cur_page += 1
-            print(self.message.author)
             await self.message.edit(embed=self.pages[self.cur_page])
 
-    async def previous_page(self):
-        if self.cur_page > 0:
+    async def previous_page(self, user_id):
+        if self.cur_page > 0 and user_id == self.owner:
             self.cur_page -= 1
             await self.message.edit(embed=self.pages[self.cur_page])
 
     async def remove(self):
-        await self.message.remove_reaction('◀', self.ctx.me)
-        await self.message.remove_reaction('⏹', self.ctx.me)
-        await self.message.remove_reaction('▶', self.ctx.me)
+        try:
+            await self.message.clear_reactions()
+        except discord.Forbidden as e:
+            await self.message.remove_reaction('◀', self.ctx.me)
+            await self.message.remove_reaction('⏹', self.ctx.me)
+            await self.message.remove_reaction('▶', self.ctx.me)
 
 
 def setup(bot: openpotd.OpenPOTD):
